refactor: restore and update Header component for improved layout
- Reintroduced the Header component in src/components/header.tsx, enhancing the layout with navigation links and a theme toggle button.
- Updated import path in layout.tsx to reflect the new Header component location, ensuring proper integration within the application.
